+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `Set` stdlib
 - `Option.force_msg`
+- Loading namespaces into the current scope (e.g. `using Pair`)
 ### Changed
 ### Removed